Added user API's for Booking, Passenger, and Payment

+++ SpringBootServer/IndiSky/src/main/java/com/indisky/entities/Passenger.java
@@ -20,7 +20,7 @@ import java.util.List;
 public class Passenger {
     @Column(name = "passenger_id")
     @Id @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Integer passengerId;
+    private long passengerId;
 
     private String fullName;
     private LocalDate dob;
